requirements.txt
annotated-types==0.7.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
PyMuPDF==1.23.1
gunicorn==23.0.0
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
preshed==3.0.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytesseract==0.3.13
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.0
scikit-learn==1.5.1
scipy==1.14.1
setuptools==74.0.0
shellingham==1.5.4
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
threadpoolctl==3.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0